Orient side LEDs on GMMK2 65% ANSI to match the rest of the matrix (#25402)
Orient side LEDs on GMMK2 65% to match the rest of the matrix

On GMMK2 P65 ANSI, the side LEDs were mapped to the wrong sides
of the keyboard resulting in effects not being consistent with
the key surface. This was most jarring in effects like pinwheel
where the pattern involves the side LEDs and there is a distinctly
different flow on the left and right halves of the board.

This PR swaps the side LEDs in the map so that they are correctly
oriented. Applied to both ANSI and ISO variants.